fixed3 burger/price/

+++ src/test/java/com/workintech/s18d1/ApplicationPropertiesAndControllerTest.java
@@ -193,7 +193,7 @@ class ApplicationPropertiesAndControllerTest {
         List<Burger> burgers = Arrays.asList(sampleBurger);
         given(burgerDao.findByPrice(sampleBurger.getPrice())).willReturn(burgers);
 
-        mockMvc.perform(get("/burger/price/{price}", sampleBurger.getPrice().intValue()))
+        mockMvc.perform(get("/burger/price/{price}", sampleBurger.getPrice()))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$", hasSize(1)))
                 .andExpect(jsonPath("$[0].name", is(sampleBurger.getName())));
